requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.3.0
gunicorn==21.2.0
